+++ drivers/gpu/nvgpu/common/fifo/tsg.c
@@ -99,7 +99,7 @@ static bool gk20a_is_channel_active(struct gk20a *g, struct channel_gk20a *ch)
 
 	for (i = 0; i < f->max_runlists; ++i) {
 		runlist = &f->runlist_info[i];
-		if (test_bit(ch->chid, runlist->active_channels)) {
+		if (test_bit((int)ch->chid, runlist->active_channels)) {
 			return true;
 		}
 	}
